New heuristics: fifo_span
Added new class (SCAlgorithms) and new method (fifo_span). This is still a rough (heuristic) method. Statistics have not been applied yet. The idea: find intervals where course(end) - course(start) is within predefined limits. Novelty in this method is the way of traversing through the input list (fifo). Imo it is good idea and it will be probably applied in future methods.