Update SCStatistics.java - added getRelativeTimes function
Linear regression: 
When working with absolute times (the time from 1970.0), the problems were noticed. Since there are always big sums and big numbers, the numeric inconsistencies were appearing and the estimated parameters were not absolutely accurate. There is a good test for this - calculating a control value as (matrix notation): trans(v)*v - (trans(f)*f + trans(n)*x'). It should be zero but it was not zero with big numbers (i.e. with absolute times). Therefore getRelativeTimes function is introduced to be used for preparing regression data. Relative times are times reduced to the first element in array (the time of the first element is 0.0). All tests pass with relative time and the estimated regression parameters are much more correct, at the end.