Updated SCAlgorithms.java - one function added
Function sieve_maxdev for finding the optimal distribution of good intervals. This function removes the need for shifting and merging intervals. The result does not depend on the direction of iterating (left-to-right or right-to-left). It produces intervals which are optimal in the sense: a) the sum of lengths is maximal b) the sum of squares of residuals (deviations) is minimal.